requirements.txt
autopep8==2.0.1
boto3==1.34.49
botocore==1.29.83
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
decorator==5.1.1
Flask==3.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
pip-review==1.3.0
psycopg2-binary==2.9.9
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-geoip-python3==1.3
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
twilio==8.12.0
ua-parser==0.16.1
urllib3==1.26.14
uWSGI==2.0.21
validators==0.20.0
Werkzeug==2.2.2
